# docs/swagger.yaml
basePath: /
definitions:
...
  model.Usuario:
    properties:
      email:
        type: string
      id:
        type: integer
      nombre:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestión de rutinas de gimnasio
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GymBro API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario con email y contraseña
      parameters:
      - description: Credenciales de login
        in: body
        name: credentials
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Credenciales inválidas
          schema:
            additionalProperties: true
            type: object
      summary: Iniciar sesión
      tags:
      - usuarios
...
  /usuarios:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de usuarios
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Usuario'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todos los usuarios
      tags:
      - usuarios
    post:
      consumes:
      - application/json
      description: Registra un nuevo usuario en el sistema
      parameters:
      - description: Datos del usuario
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/model.Usuario'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Usuario'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Registrar nuevo usuario
      tags:
      - usuarios
  /usuarios/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un usuario del sistema
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuario eliminado correctamente
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar usuario
      tags:
      - usuarios
    get:
      consumes:
      - application/json
      description: Obtiene un usuario específico por su ID
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "404":
          description: Usuario no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener usuario por ID
      tags:
      - usuarios
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un usuario existente
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del usuario
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/model.Usuario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Usuario no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar usuario
      tags:
      - usuarios
  /usuarios/email/{email}:
    get:
      consumes:
      - application/json
      description: Obtiene un usuario específico por su email
      parameters:
      - description: Email del usuario
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "404":
          description: Usuario no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener usuario por email
      tags:
      - usuarios
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"